refactor(app): move content-progress store out of globalConfig
