+++ app/src/components/content/AudioPlayer.vue
@@ -21,15 +21,8 @@ import {
 } from "@heroicons/vue/20/solid";
 import LImage from "@/components/images/LImage.vue";
 import { DateTime } from "luxon";
-import {
-    clearMediaQueue,
-    cmsLanguages,
-    getMediaProgress,
-    isMobileScreen,
-    mediaQueue,
-    removeMediaProgress,
-    setMediaProgress,
-} from "@/globalConfig";
+import { clearMediaQueue, cmsLanguages, isMobileScreen, mediaQueue } from "@/globalConfig";
+import { getMediaProgress, removeMediaProgress, setMediaProgress } from "@/contentProgress";
 import LDialog from "@/components/common/LDialog.vue";
 
 const router = useRouter();
@@ -1240,7 +1233,7 @@ watch(matchAudioFileUrl, async (newUrl, oldUrl) => {
                         <!-- Volume toggle button -->
                         <button
                             @click="toggleVolumeSlider"
-                            class="flex min-h-[44px] min-w-[44px] touch-manipulation items-center justify-center space-x-1 rounded-lg px-3 py-2 hover:bg-black/10 dark:hover:bg-white/10 active:bg-black/10 dark:active:bg-white/10"
+                            class="flex min-h-[44px] min-w-[44px] touch-manipulation items-center justify-center space-x-1 rounded-lg px-3 py-2 hover:bg-black/10 active:bg-black/10 dark:hover:bg-white/10 dark:active:bg-white/10"
                             :title="isMuted ? 'Unmute (M)' : 'Mute (M)'"
                             aria-label="Toggle volume controls"
                         >
@@ -1446,7 +1439,7 @@ watch(matchAudioFileUrl, async (newUrl, oldUrl) => {
                 <div class="volume-control-container relative hidden lg:block">
                     <button
                         @click.stop="toggleVolumeSlider"
-                        class="flex h-11 w-11 flex-shrink-0 touch-manipulation items-center justify-center rounded-full bg-transparent hover:bg-black/10 dark:hover:bg-white/10 active:bg-black/10 dark:active:bg-white/10"
+                        class="flex h-11 w-11 flex-shrink-0 touch-manipulation items-center justify-center rounded-full bg-transparent hover:bg-black/10 active:bg-black/10 dark:hover:bg-white/10 dark:active:bg-white/10"
                         title="Volume controls"
                         aria-label="Toggle volume controls"
                     >
@@ -1755,7 +1748,6 @@ watch(matchAudioFileUrl, async (newUrl, oldUrl) => {
         transform 0.3s ease-out,
         opacity 0.3s ease-out;
 }
-
 
 .slide-up-enter-from,
 .slide-up-leave-to {
